make tracked AuthorHead default export
